move forever impl to game
