+++ libs/base/basic.ts
@@ -1,56 +1,1 @@
-/**
- * Busy wait for a condition to be true
- * @param condition condition to test for
- * @param timeOut if positive, maximum duration to wait for in milliseconds
- */
-//% blockId="pxt_pause_until"
-function pauseUntil(condition: () => boolean, timeOut?: number): void {
-    if (!condition || condition()) return; // optimistic path
-    if (!timeOut) timeOut = 0;
-    control.__queuePollEvent(timeOut, condition, undefined);
-}
-
-/**
- * Repeats the code forever in the background. On each iteration, allows other codes to run.
- * @param body code to execute
- */
-//% help=loops/forever weight=100 afterOnStart=true blockNamespace="loops"
-//% blockId=forever block="forever" blockAllowMultiple=1
-function forever(a: () => void): void {
-    game.forever(a);
-}
-
-/**
- * Pause for the specified time in milliseconds
- * @param ms how long to pause for, eg: 100, 200, 500, 1000, 2000
- */
-//% help=loops/pause weight=99
-//% async block="pause %pause=timePicker|ms"
-//% blockId=device_pause blockNamespace="loops"
-function pause(ms: number): void {
-    loops.pause(ms);
-}
-
-/**
- * Convert any value to text
- * @param value value to be converted to text
- */
-//% help=text/convert-to-text weight=1
-//% block="convert $value=math_number to text"
-//% blockId=variable_to_text blockNamespace="text"
-function convertToText(value: any): string {
-    return "" + value;
-}
-
-// micro:bit compatibility
-// these functions allow some level of reuse
-// between micro:bit and other maker-style editors
-namespace basic {
-    export function pause(millis: number) {
-        loops.pause(millis);
-    }
-
-    export function forever(a: () => void) {
-        game.forever(a);
-    }
-}
+// implemented in game
